// src/types/config.ts
export interface ConfigData {
  id?: number;
  key: string;
  value: string;
  description?: string;
}